moved notifications into events api / tweaked ordering
